refactor: derive account into helper
